Made changes
I worked with my TL to fix some issues like, the data wasn't matching with the back-end in the SignUp.jsand same with the SignIn.js, both are resolved. I took out redux because we aren't going to use it anymore. I also got a new plan for us all that I came up with my TL, this is going to help me bring together the App to make it all connect to one another! =)