Updates
We were able to have a TL come in and try to help us with the issue we were having, it has been resolved for most parts, now waiting for backend to come back from stand up to help set things up correctly. Made minor changes due to the TL for testing.